Speed up clustering test suite (~7x)
Profiling ddmc/tests/test_cluster.py showed most time was spent in
overhead unrelated to the actual EM algorithm:

- Binomial's background PWM went through Biopython's general-purpose
  Bio.Align/Bio.motifs machinery; replaced with a direct numpy
  implementation (fast_position_weight_matrix), verified numerically
  identical.
- Binomial.from_summaries ran two np.einsum calls per EM iteration on
  a boolean one-hot array; replaced with flattened float32 matmuls.
- BackgroundSeqs re-read and re-filtered the ~360k-row PhosphoSitePlus
  CSV from scratch on every Binomial init; the reference data and the
  per-proportion background sequence list are now cached with
  lru_cache.
- PAM250 re-cast its int8 pairwise-distance matrix to float64 on every
  EM iteration instead of once at init.
- Every test independently reloaded and refiltered the same CPTAC
  p-signal CSV; consolidated into a session-scoped pytest fixture,
  since DDMC.fit doesn't mutate its input when there's no missing data.

Full suite time: 38.5s -> 5.2s.

+++ ddmc/pam250.py
@@ -6,8 +6,10 @@ from Bio.Align import substitution_matrices
 
 class PAM250:
     def __init__(self, seqs: list[str]):
-        # Compute all pairwise distances
-        self.background = get_pam250_scores(seqs)
+        # Compute all pairwise distances. Cast to float32 once here rather
+        # than in from_summaries, which runs every EM iteration and would
+        # otherwise re-convert this (potentially large) int8 matrix each time.
+        self.background = get_pam250_scores(seqs).astype(np.float32)
         self.logWeights = 0.0
 
     def from_summaries(self, weightsIn: np.ndarray):
